requirements.txt
obspy==1.3.0
google-auth==2.11.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.65.0
requests==2.28.1
